reorganized solver mapping types, currently broken animate function
