+++ src/Store.tsx
@@ -4,7 +4,7 @@ import Vec2 from "./helpers/Vec2";
 import Node from "./helpers/Node";
 import type { NodeType } from "./@types/helpers/Node";
 import { AlgorithmParams, Generic } from "./@types/helpers/Node";
-import type { Algorithm } from "./@types/components/AlgorithmMenu";
+import type { Algorithm } from "./@types/helpers/algorithms";
 
 export const StoreContext = createContext<StoreContextType | null>(null);
 
@@ -16,7 +16,7 @@ const StoreProvider : React.FC<StoreProviderProps> = ({ children }) => {
   const [startIdx, setStartIdx] = useState<Vec2 | undefined>()
   const [targetIdx, setTargetIdx] = useState<Vec2 | undefined>();
   const [isStarted, setIsStarted] = useState<boolean>(false);
-  const [selectedAlgorithm, setSelectedAlgorithm] = useState<Algorithm | undefined>();
+  const [selectedAlgorithm, setSelectedAlgorithm] = useState<Algorithm>("astar");
   // const [state, setState] = useState<StoreContextType>({ 
   //   cellSize: undefined
   // })
