+++ src/utils/types.ts
@@ -1,4 +1,4 @@
-import type { Algorithm } from "../@types/components/AlgorithmMenu";
+import type { Algorithm } from "../@types/helpers/algorithms"
 
 export const isAlgorithm = (val: string) : val is Algorithm => {
   return ["astar", "djikstra"].includes(val)
